i1577 - bug fix
