+++ src/main/scala/scorex/core/utils/utils.scala
@@ -1,8 +1,8 @@
 package scorex.core
 
 import java.security.SecureRandom
-
 import scala.annotation.tailrec
+import scala.collection.mutable
 import scala.concurrent.duration._
 import scala.util.{Failure, Success, Try}
 
@@ -74,4 +74,27 @@ package object utils {
     result
   }
 
+
+  implicit class MapPimp[K, V](underlying: mutable.Map[K, V]) {
+    /**
+      * One liner for updating a Map with the possibility to handle case of missing Key
+      * @param k map key
+      * @param f function that is passed Option depending on Key being present or missing, returning new Value
+      * @return Option depending on map being updated or not
+      */
+    def adjust(k: K)(f: Option[V] => V): Option[V] = underlying.put(k, f(underlying.get(k)))
+
+    /**
+      * One liner for updating a Map with the possibility to handle case of missing Key
+      * @param k map key
+      * @param f function that is passed Option depending on Key being present or missing,
+      *          returning Option signaling whether to update or not
+      * @return new Map with value updated under given key
+      */
+    def flatAdjust(k: K)(f: Option[V] => Option[V]): Option[V] =
+      f(underlying.get(k)) match {
+        case None    => None
+        case Some(v) => underlying.put(k, v)
+      }
+  }
 }
